Add block patterns e2e tests (#22414)
