+++ packages/e2e-test-utils/README.md
@@ -288,6 +288,15 @@ _Parameters_
 
 -   _searchTerm_ `string`: The text to search the inserter for.
 
+<a name="insertPattern" href="#insertPattern">#</a> **insertPattern**
+
+Opens the inserter, searches for the given pattern, then selects the first
+result that appears. It then waits briefly for the block list to update.
+
+_Parameters_
+
+-   _searchTerm_ `string`: The text to search the inserter for.
+
 <a name="installPlugin" href="#installPlugin">#</a> **installPlugin**
 
 Installs a plugin from the WP.org repository.
@@ -407,6 +416,14 @@ _Returns_
 <a name="searchForBlock" href="#searchForBlock">#</a> **searchForBlock**
 
 Search for block in the global inserter
+
+_Parameters_
+
+-   _searchTerm_ `string`: The text to search the inserter for.
+
+<a name="searchForPattern" href="#searchForPattern">#</a> **searchForPattern**
+
+Search for pattern in the global inserter
 
 _Parameters_
