+++ packages/e2e-test-utils/src/index.js
@@ -23,7 +23,14 @@ export { getBlockSetting } from './get-block-setting';
 export { getEditedPostContent } from './get-edited-post-content';
 export { hasBlockSwitcher } from './has-block-switcher';
 export { getPageError } from './get-page-error';
-export { insertBlock } from './insert-block';
+export {
+	insertBlock,
+	insertPattern,
+	searchForBlock,
+	searchForPattern,
+	openGlobalBlockInserter,
+	closeGlobalBlockInserter,
+} from './inserter';
 export { installPlugin } from './install-plugin';
 export { isCurrentURL } from './is-current-url';
 export { isInDefaultBlock } from './is-in-default-block';
@@ -33,17 +40,12 @@ export {
 	disableFocusLossObservation,
 } from './observe-focus-loss';
 export { openDocumentSettingsSidebar } from './open-document-settings-sidebar';
-export {
-	openGlobalBlockInserter,
-	closeGlobalBlockInserter,
-} from './open-global-block-inserter';
 export { openPublishPanel } from './open-publish-panel';
 export { pressKeyTimes } from './press-key-times';
 export { pressKeyWithModifier } from './press-key-with-modifier';
 export { publishPost } from './publish-post';
 export { publishPostWithPrePublishChecksDisabled } from './publish-post-with-pre-publish-checks-disabled';
 export { saveDraft } from './save-draft';
-export { searchForBlock } from './search-for-block';
 export { selectBlockByClientId } from './select-block-by-client-id';
 export { setBrowserViewport } from './set-browser-viewport';
 export { setPostContent } from './set-post-content';
