--- packages/e2e-test-utils/src/insert-block.js
@@ -1,18 +1,0 @@
-/**
- * Internal dependencies
- */
-import { searchForBlock } from './search-for-block';
-
-/**
- * Opens the inserter, searches for the given term, then selects the first
- * result that appears. It then waits briefly for the block list to update.
- *
- * @param {string} searchTerm The text to search the inserter for.
- */
-export async function insertBlock( searchTerm ) {
-	await searchForBlock( searchTerm );
-	const insertButton = (
-		await page.$x( `//button//span[contains(text(), '${ searchTerm }')]` )
-	 )[ 0 ];
-	await insertButton.click();
-}
